Use new CalendarPageLoader in EphemerisCalendar.kt

+++ android/views/src/main/kotlin/com/boswelja/ephemeris/views/CalendarPagerAdapter.kt
@@ -24,7 +24,7 @@ internal class CalendarPagerAdapter(
     }
 
     override fun onBindViewHolder(holder: CalendarPageViewHolder, position: Int) {
-        val page = (position - (Int.MAX_VALUE / 2)).toLong()
+        val page = (position - (Int.MAX_VALUE / 2))
         val pageState = pagingSource.loadPageData(page) { date, month ->
             DisplayDate(date, focusMode(date, month))
         }
